Fix bug when user scrolls content quickly. Closes #37

+++ app/src/main/java/saulmm/myapplication/AvatarImageBehavior.java
@@ -85,7 +85,7 @@ public class AvatarImageBehavior extends CoordinatorLayout.Behavior<CircleImageV
 
             child.setX(mStartXPosition - distanceXToSubtract);
             
-            int maxYDistance = (mStartYPosition - mFinalYPosition) + (mCustomFinalHeight / 2);
+            float maxYDistance = (mStartYPosition - mFinalYPosition) + (mCustomFinalHeight / 2);
             if (distanceYToSubtract > maxYDistance)
                 distanceYToSubtract = maxYDistance;
